Ecore moving to dedicated folder
