+++ main/src/test/resources/testCases/validation/ecore/broken-dangling-reference.ecore
