+++ main/src/test/resources/testCases/validation/ecore/broken-malformed-xml.ecore
