+++ main/src/test/resources/testCases/validation/ecore/broken-uncompilable.ecore
