+++ main/src/test/resources/testCases/validation/ecore/valid.ecore
